extract Tweet Box to be a separate component
