+++ client/src/components/Tweet/Tweet.js
@@ -6,50 +6,8 @@ import { Icon } from "react-icons-kit";
 import { repeat } from "react-icons-kit/feather/repeat";
 import { CurrentUserContext } from "../CurrentUserContext";
 import TweetActions from "./TweetActions";
+import TweetBox from "./TweetBox";
 
-const TweetBoxWrapper = styled.form`
-  margin-top: 40px;
-  border-bottom: 10px solid #00000017;
-  width: 500px;
-  display: flex;
-  flex-direction: column;
-  padding: 10px;
-`;
-
-const TweetBoxHeader = styled.div`
-  display: flex;
-`;
-const TweetBoxInput = styled.textarea`
-  width: 300px;
-  height: 50px;
-  overflow-x: hidden;
-  overflow-y: scroll;
-  scrollbar-color: grey;
-  margin-left: 10px;
-  resize: none;
-  border: none;
-  outline: none;
-  font-size: 16px;
-`;
-
-const UserAvatar = styled.img`
-  border-radius: 50%;
-  width: 50px;
-  height: 50px;
-`;
-
-const MeowCTA = styled.input`
-  background: hsl(258deg, 100%, 50%);
-  color: white;
-  width: 100px;
-  border-radius: 35px;
-  height: 30px;
-  font-weight: bold;
-  cursor: pointer;
-  outline: none;
-  border: none;
-  align-self: flex-end;
-`;
 const TweetFeedWrapper = styled.div`
   margin-top: 20px;
   display: grid;
@@ -130,99 +88,18 @@ const StyledLink = styled(Link)`
   color: inherit;
 `;
 
-const CharacterCount = styled.p`
-  color: ${props => props.charColor};
-`;
-
 export default function Tweet() {
   const {
     currentUser,
     userStatus,
     feed,
     feedStatus,
-    setFeedStatus
+    setFeedStatus,
   } = React.useContext(CurrentUserContext);
-
-  const [inputText, setInputText] = useState("");
-
-  const [charCount, setCharCount] = useState({ current: 0, maxChar: 280 });
-  //   const [charColor, setCharColor] = useState("black");
-
-  const inputEl = useRef(null);
-
-  const isDisabled = useMemo(() => {
-    if (inputText.length <= 280) {
-      return false;
-    }
-    return true;
-  }, [inputText]);
-
-  const handleSubmit = e => {
-    e.preventDefault();
-    if (inputText.length <= 280) {
-      fetch("/api/tweet", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ status: inputText })
-      })
-        .then(res => res.json())
-        .then(res => {
-          if (res) {
-            setFeedStatus("Loading");
-            //   console.log(res);
-          }
-        });
-    } else {
-      setFeedStatus("Loading");
-    }
-  };
-
-  const handleChange = event => {
-    setInputText(event.target.value);
-    setCharCount({
-      ...charCount,
-      current: event.target.value.length
-    });
-  };
-
-  //   console.log(charCount.maxChar - charCount.current);
-  useEffect(() => {
-    if (userStatus === "ok" && feedStatus === "ok") {
-      inputEl.current.focus();
-    }
-  }, [userStatus, feedStatus]);
 
   return (
     <>
-      {userStatus === "ok" && feedStatus === "ok" && (
-        <TweetBoxWrapper onSubmit={handleSubmit}>
-          {/* {console.log(currentUser)} */}
-          <TweetBoxHeader>
-            <UserAvatar src={currentUser.avatarSrc} />
-            <TweetBoxInput
-              ref={inputEl}
-              placeholder={"What's happening?"}
-              onChange={handleChange}
-            ></TweetBoxInput>
-          </TweetBoxHeader>
-          {charCount.maxChar - charCount.current <= 55 &&
-          charCount.maxChar - charCount.current >= 0 ? (
-            <CharacterCount charColor={"#f0e130"}>
-              {charCount.maxChar - charCount.current}
-            </CharacterCount>
-          ) : charCount.maxChar - charCount.current < 0 ? (
-            <CharacterCount charColor={"red"}>
-              {charCount.maxChar - charCount.current}
-            </CharacterCount>
-          ) : (
-            <CharacterCount charColor={"black"}>
-              {charCount.maxChar - charCount.current}
-            </CharacterCount>
-          )}
-
-          <MeowCTA type="submit" value={"MEOW"} disabled={isDisabled}></MeowCTA>
-        </TweetBoxWrapper>
-      )}
+      <TweetBox />
       <TweetFeedWrapper>
         {userStatus === "ok" && feedStatus === "ok" ? (
           feed.tweetIds.map((tweet, i) => {
